Refactor MovieCard location
